Generate only one full-courses.json (#458)
