+++ package.json
@@ -5,7 +5,6 @@
   "scripts": {
     "serve": "vue-cli-service serve",
     "build": "vue-cli-service build --mode production",
-    "fetch-courses": "ts-node -T -P tsconfig.node.json src/requirements/fetcher.ts",
     "courses-gen": "ts-node -T -P tsconfig.node.json src/requirements/courses-json-generator.ts",
     "req-gen": "ts-node -T -P tsconfig.node.json src/requirements/requirement-json-generator.ts",
     "lint": "vue-cli-service lint",
