+++ src/requirements/requirement-json-generator.ts
@@ -7,7 +7,7 @@ import {
 } from './types';
 import sourceRequirements from './data';
 import { FWS_COURSE_ID, CREDITS_COURSE_ID } from './data/constants';
-import { coursesWithLastOfferingData } from './filtered-course-data';
+import { fullCoursesArray } from '../assets/courses/typed-full-courses';
 
 /**
  * Special (synthetic) courses, as used in AP/IB equivalent courses generation.
@@ -38,7 +38,7 @@ const getEligibleCoursesFromRequirementCheckers = (
 ): readonly (readonly number[])[] =>
   checkers.map(oneRequirementChecker => {
     const courseIdSet = new Set(
-      [...coursesWithLastOfferingData, ...specialCourses]
+      [...fullCoursesArray, ...specialCourses]
         .filter(course => oneRequirementChecker(course))
         .map(course => course.crseId)
     );
