+++ src/requirements/types.ts
@@ -1,21 +1,4 @@
-export type Course = {
-  readonly subject: string;
-  readonly crseId: number;
-  readonly catalogNbr: string;
-  readonly titleLong: string;
-  readonly enrollGroups: readonly {
-    readonly unitsMinimum: number;
-    readonly unitsMaximum: number;
-  }[];
-  readonly catalogWhenOffered?: string;
-  readonly catalogBreadth?: string;
-  readonly catalogDistr?: string;
-  readonly catalogComments?: string;
-  readonly catalogSatisfiesReq?: string;
-  readonly catalogCourseSubfield?: string;
-  readonly acadCareer: string;
-  readonly acadGroup: string;
-};
+export type Course = Omit<CornellCourseRosterCourse, 'roster'>;
 
 export type BaseRequirement = RequirementCommon & RequirementFulfillmentInformation;
